sales_docs_to_podcast/requirements.txt
elevenlabs==2.16.0
python-dotenv==1.1.1
openai==1.109.1
requests==2.32.5